requirements.txt
asgiref==3.6.0
certifi==2022.12.7
Django==4.2
django-elasticsearch-dsl==7.3
djangorestframework==3.14.0
elasticsearch==7.17.9
elasticsearch-dsl==7.4.1
pkg_resources==0.0.0
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
urllib3==1.26.15
